Various Improvements
Change flags to mini armor stands with wool on their heads so that they
are visible from a far distance.
Make the flags visible in the start game state so players can see the
flag before the game starts.
Add particles that signify each team's flag spawn. Especially visible
when the flag is not there.
Change default winning score to 5.
Add a "rebalance teams" option in the control panel book.
Add an "Advanced Controls" page to the control panel book for finer
control over who gets added to which team, for manually setting a player
to spectate, and have finer control of the time limit and winning score.